backend/requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.34.3
sqlalchemy[asyncio]==2.0.48
asyncpg==0.30.0
geoalchemy2==0.15.2
alembic==1.14.1
pydantic-settings==2.7.1
supabase==2.28.3
python-jose[cryptography]==3.3.0
httpx==0.28.1
